write db_metadata_nexus records with the correct state
